change mouse sketch functionality
Previously, when the user hovered over the canvas it triggered a
mouseover event, automatically causing the user to sketch by filling in
the background color. This was a bit annoying as it didn't give the user
very much control. Now, the sketch function will only fire while the
mouse is held down.